requirements.txt
# requirements.txt
PyQt5
Pillow
pytesseract
